requirements.txt
django==1.4.5
pillow==1.7.7
south==0.7.6
whoosh==2.4.1
pystemmer==1.2.0
django-guardian==1.0.4
django-mailer==0.1.0
django-mptt==0.5.5
django-widget-tweaks==1.1.2
django-simple-captcha==0.4.2
